+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 ### Added:
+- Added `submit_and_wait` to sign (if needed), autofill, submit a transaction and wait for its final outcome
+- `submit` and `send_reliable_submission` now accept an optional boolean param `fail_hard` (if `True` halt the submission if it's not immediately validated)
 - Support for cross-chain bridge proposal
 - Support for Automated Market Maker (AMM) transactions and requests as defined in XLS-30.
 
 ### Fixed:
-- Refactored `does_account_exist` to avoid deprecated methods and use `ledger_index` parameter
+- Refactored `does_account_exist` and `get_balance` to avoid deprecated methods and use `ledger_index` parameter
+
+### Removed:
+- RPCs and utils related to the old sidechain design
 
 ## [1.8.0] - 2023-04-10
 ### Added: